+++ examples/long-polling.js
@@ -5,7 +5,7 @@ const client = new ZetaPush.WeakClient({
 })
 // Create a Macro service
 const service = client.createService({
-  type: ZetaPush.services.Macro,
+  Type: ZetaPush.services.Macro,
   listener: {
     error({ data }) {
       console.error('macro error', data)
@@ -28,7 +28,7 @@ document.querySelector('.js-SayHello').addEventListener('click', () => {
   service.call({
     name: 'hello',
     parameters: {
-      value: 'World'
+      name: 'World'
     }
   })
 })