+++ examples/messaging.js
@@ -4,7 +4,7 @@ const client = new ZetaPush.WeakClient({
 })
 // Create Messaging service
 const service = client.createService({
-  type: ZetaPush.services.Messaging,
+  Type: ZetaPush.services.Messaging,
   listener: {
     error(error) {
       console.log('error', error)
@@ -27,7 +27,7 @@ document.querySelector('.js-SendMessage').addEventListener('click', () => {
   service.send({
     target: prompt('Target User Id', client.getUserId()),
     data: {
-      value: 'World'
+      name: 'World'
     }
   })
 })