+++ examples/unsubscribe.js
@@ -2,7 +2,7 @@ const client = new ZetaPush.WeakClient({
   sandboxId: 'Y1k3xBDc'
 })
 const service = client.createService({
-  type: ZetaPush.services.Macro,
+  Type: ZetaPush.services.Macro,
   listener: {
     error(message) {
       console.error('macro error', message.data)
@@ -23,7 +23,7 @@ document.querySelector('.js-SayHello').addEventListener('click', () => {
   service.call({
     name: 'hello',
     parameters: {
-      value: 'World'
+      name: 'World'
     }
   })
 })