+++ CHANGELOG.md
@@ -1,3 +1,102 @@
+# 3.0.0 (2016-10-24)
+
+### Changes
+
+* **client** Rename createService parameter from type to Type
+* **core:** Update lib/client, rename credentials to authentication
+
+### Fixes
+
+* **smart-client** Fix sessions persistence key generation
+
+### Features
+
+* **core** Add lib/smart-client
+* **test** ZetaPush core functionalities are now tested with Karma/Jasmine
+
+# 3.0.0-alpha.3 (2016-09-16)
+
+Major API modification
+
+### BREAKING CHANGES
+
+* **client** Rename createService parameter from type to Type
+
+**Before**
+
+```js
+// Create new service
+const service = new client.createService({
+  type: ZetaPush.services.Stack,
+  listener: {
+    list(message) {
+      console.log('on list', message)
+    }
+  }
+})
+```
+
+**After**
+
+```js
+// Create new service
+const service = new client.createService({
+  Type: ZetaPush.services.Stack,
+  listener: {
+    list(message) {
+      console.log('on list', message)
+    }
+  }
+})
+```
+
+# 3.0.0-alpha.2 (2016-09-06)
+
+### Fix
+
+* **smart-client** Fix sessions persistence key generation
+
+# 3.0.0-alpha.1 (2016-08-19)
+
+Major API modification
+
+### BREAKING CHANGES
+
+* **core:** Update lib/client, rename credentials to authentication
+
+**Before**
+
+```js
+// Create new ZetaPush Client
+const client = new ZetaPush.Client({
+  sandboxId: '<YOUR-SANDBOX-ID>',
+  credentials() {
+    return ZetaPush.AuthentFactory.createWeakHandshake({
+      token: null
+    })
+  }
+})
+```
+
+**After**
+
+```js
+// Create new ZetaPush Client
+const client = new ZetaPush.Client({
+  sandboxId: '<YOUR-SANDBOX-ID>',
+  authentication() {
+    return ZetaPush.Authentication.weak({
+      token: null
+    })
+  }
+})
+```
+
+### Features
+
+* **core** Add lib/smart-client
+* **test** ZetaPush core functionalities are now tested with Karma/Jasmine
+
 # 2.0.1 (2016-07-26)
 
 ### Fixes